module: associate KernelModule with drgn.Module
Drgn 0.0.31 introduces the module API, and every kernel module will have
an automatically-detected drgn.Module associated with it by drgn. This
Module is the key to seeing information about whether debuginfo is
loaded, and it also has efficient lookup APIs. Remove duplicate and
unused functionality drgn_tools.module and rely on drgn.Module for what
we need.

Some functionality, like get_symbol(), is now totally unused. That code
also relates to the ModuleSymbolFinder, which is similarly unused, and
even if it were, it could be replaced by the SymbolIndex which I've
upstreamed to drgn. The only symbol-releated functionality which still
belongs is module_exports() / KernelModule.exports().

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>