corelens: suppress drgn's MissingDebugInfo warning
Drgn's logging, in general, is quite useful, and in the future we may
include a way to control the log level used by corelens & drgn. However,
drgn logs a warning for missing debug files when you use the
load_default_debug_info() function, which ruins the corelens output. We
really don't want this, because corelens already handles module
debuginfo. Add a filter to the drgn logger in order to handle this.

While we're at it, we can simplify our logging in Corelens by quite a
bit, by using the logging module instead of the existing bespoke system.
Thankfully, logging has all the handler components that are necessary to
easily reproduce the same behavior.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>