debuginfo: handle cold-patches and build-id mismatches
In the happy case, every kernel module which is not OOT-tainted will be
found in a kernel RPM. This is true in the overwhelming majority of
cases, but a few edge cases can cause undesirable behavior.

1. Ksplice cold-patch modules are expected to be mismatched. The
shipped debuginfo doesn't apply for them.
2. Some old kernels had module build ID mismatches which cannot be
corrected.

In either case, drgn will reject the debuginfo and continue calling
debuginfo finders. This can result in our finders being called multiple
times, pointlessly downloading and/or extracting RPMs multiple times.

To handle cold-patches, we just don't consider them in-tree modules. To
handle the remaining build ID mismatches, we record all module names for
which we've attempted RPM extraction. If we're later called, we check
each module against that set and avoid re-doing the work.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>