debuginfo: keep downloaded RPM until exit
Debug info searches go in two steps: (1) kernel, (2) modules. You cannot
know which modules must be loaded until the kernel is loaded. This is a
problem for RPM downloading and extraction. While we can't avoid the
double-extraction, it is pretty wasteful to download the RPM once for
each step.

The rpm_cache configuration was originally conceived as a band-aid over
this issue. It lets the RPM get saved persistently, but it has the nice
upside that the ol-local-rpm would find the saved RPM and ensure that
the ol-download finder doesn't get called again. That's nice, but in
reality we probably don't want rpm_cache enabled all that often.
Instead, persist the downloaded file until the program exits. Ensure
that even if ol-local-rpm is disabled, the ol-download finder will still
call it with the cached RPM as necessary.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>