list_lru: further limit test output and duration
The "maxitems" parameter set at 10,000 is tracked per mount, which means
that the list_lru test can output far more than that. Drop this down to
100 per mount, only allow the test to process 20 mounts before
completing.

The purpose of the smoke test is just to verify that we have iterated
over the data structures correctly. Having small limits is fine for that
purpose. Custom (or zero) amounts can still be specified for those
interested in running the full operation.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>