pstack: import from drgn
The pstack script allows dumping userspace stack traces for processes.
It has a dump mode which only dumps minimal necessary information,
suitable for use within a kdump kernel. Finally, it has a "print" mode
to print the dumped contents. Import this script and make a small
CorelensModule wrapper around it. We'll refine some of the functionality
so that it can be supported in LKCE.

The drgn file contrib/pstack.py was wholely contributed by me under
drgn's license. As the author, I can also add it here under the
UPL.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>