pstack: enable unwinding running tasks
In most vmcores, userspace tasks are not actually running: the CPU is
executing a kernel-mode interrupt handler that brings it to a halt.
Thus, the userspace registers are stored to the kernel stack. But for
hypervisor vmcores, the CPU may be executing user code directly. In that
case, drgn's stack trace registers are in fact the userspace registers.
Handle this (rare) case.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>