pstack: dump all data to a single file
The original pstack implementation dumps data to multiple files: for
each process, created one dump file for the JSON metadata, and another for
the stack data. For large amounts of tasks, this can result in a lot of
file creation, which can be unwieldy, and it's probably not a great idea
to create thousands of tiny files in the kdump kernel.

Move to a new format where everything is dumped to a single file. This
requires a bit more code and a clear format definition, but the result
is more efficient and easier to use.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>