pstack: make task selection much more flexible
Previously, task selection could only be done with one CLI option. So a
single PID could be specified, or a single command pattern, or a single
task state. This is obviously not flexible enough. We would like to be
able to select multiple patterns where possible. Implement this while
ensuring we don't dump or print any tasks twice.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>