conftest: allow setting a minimum kernel version
For a while we've been beating around the bush by having a way to filter
out which vmcores we run tests against. But that doesn't help filtering
out kernel versions for live tests.

Most new functionality doesn't necessarily need to work on UEK4, so we
want it to be easy to skip tests below the minimum version.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>