debuginfo: stop trying to detect cold-patches
Ksplice cold-patches are a pain because they are kernel modules whose
build IDs (and debuginfo) are mismatched, but they otherwise look just
like the in-tree module.

Previously we tried to detect them, and thus avoid attempting to extract
or load their debuginfo. However, in practice this doesn't seem
feasible. While I've been able to find some signals that a module may be
a cold-patch, none have generalized to all architectures and versions.

Instead, we need to just handle the effects of this problem. When
cold-patches aren't handled, the ol-download and ol-local-rpm finders
will repeat attempting to download & extract these debuginfo files,
every time they're used. We already have some safeguards to prevent
double-execution (download, then re-extract). But we can extend this
safeguard to the case where we've previously extracted the RPM. If we
already tried the file from the vmlinux_repo, then there's no point in
trying to download or extract that module again.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>