debuginfo: report the debuginfo file for DWARF & CTF
With the module API we can report the actual DWARF file that gets
loaded. But CTF wasn't explicitly reported. Given that the Oracle plugin
now handles CTF loading, we can also save the file that got loaded, so
that we can later report it for the CLI or corelens logs.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>