debuginfo: also handle mismatches with debuginfo RPM
While a recent commit handled the case where we had extracted files from
a downloaded RPM, and there was a build ID mismatch, there was still the
case where the debuginfo RPM was installed to the system. Since drgn's
standard finder loads those files, we would never have the opportunity
to populate the "extracted" set for those modules. Thus, when the
debuginfo RPM is installed, it would be possible for us to try to
download and extract debuginfo in the presence of a build ID
mismatch (e.g. ksplice cold-patch). Avoid this and also report a warning.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>