rds: eliminate outfile & report parameters
Maintaining the outfile & report parameters is a bit difficult for a few
reasons. First, the "outfile" parameter is a string filename, which
means that whenever an output must be written, the file must be opened.
Second, the "report" parameter is intended to determine the mode (append
vs write), but this becomes less than useful if you need to write
multiple things at a time: when report is False, you'll only get the
last item printed.

The intended use case for these parameters seems to be so that we can
easily provide custom RDS scripts to customers. The idea being that many
outputs would be too large, so we may need to only run certain
functions, and redirect output to several files for ease of access.

To support this, let's create a @redirectable decorator. It will take
any function, and allow it to accept an "outfile" parameter. When
provided, this parameter will redirect the function's output to the
file. An optional :w or :a can be appended to the filename in order to
specify the mode (it is :w by default). All print statements can simply
write to stdout, and it will be redirected appropriately where
necessary. For example, a custom script could now be created easily:

    from drgn_tools import rds
    rds.rds_conn_info(prog, outfile="conn_info.txt")
    rds.rds_sock_info(prog, outfile="other_data.txt:a")
    rds.rdma_resource_usage(prog, outfile="other_data.txt:a")

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>