testing.litevm: skip up to 4 kernels w/o debuginfo
We've seen tests fail due to debuginfo being missing from
oss.oracle.com. This is an ongoing issue and rather than have it block
our tests, we need to add a bit more flexibility.

This is still only activated when the github label is applied to the PR,
so in most cases, up-to-date debuginfo is still required. This just
unblocks pull request review and merge.

This commit also includes the ability to freeze tests at a specific
version, which is useful for keeping tests passing on the stable
branches.

Orabug: 38625030

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>